feat/staff-total: Still tinkering iwth shared component...
