+++ src/app/features/workplace/total-staff-question/total-staff-question.component.html
@@ -1,8 +1,7 @@
-<form #formEl novalidate (ngSubmit)="onSubmit()" [formGroup]="totalStaffComponent.form">
+<form #formEl novalidate (ngSubmit)="onSubmit()" [formGroup]="form">
   <app-total-staff
     [establishmentUid]="establishment.uid"
-    [form]="form"
-    [formErrorsMap]="formErrorsMap">
+    [form]="form">
   </app-total-staff>
 
   <app-submit-button [return]="!!return" (clicked)="onSubmit($event)"></app-submit-button>
